requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.35
asyncpg==0.30.0
alembic==1.14.0
redis[hiredis]==5.2.0
httpx==0.28.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic==2.10.0
pydantic-settings==2.6.0
Pillow==11.1.0
stripe==11.0.0
python-multipart==0.0.12
structlog==24.4.0
pytest==8.3.0
pytest-asyncio==0.24.0
pytest-cov==6.0.0
pip-audit
